Working over some code
