+++ football/app/model/model.scala
@@ -1,7 +1,7 @@
 package model
-import pa.{ LeagueTableEntry, FootballMatch, Result, Fixture }
+import pa._
 import org.joda.time.DateMidnight
-import feed._
+import pa.LeagueTableEntry
 
 case class Competition(
     id: String,
@@ -17,4 +17,20 @@ case class Competition(
   lazy val hasLeagueTable = leagueTable.nonEmpty
 
   lazy val matchDates = matches.map(_.date.toDateMidnight).distinct
+}
+
+case class Group(round: Option[Round], entries: Seq[LeagueTableEntry])
+
+case class Table(competition: Competition, groups: Seq[Group]) {
+  lazy val multiGroup = groups.size > 1
+}
+
+object Table {
+  def apply(competition: Competition): Table = {
+    val groups = competition.leagueTable
+      .groupBy(_.round)
+      .map { case (round, table) => Group(round, table) }
+      .toSeq.sortBy(_.round.map(_.roundNumber).getOrElse(""))
+    Table(competition, groups)
+  }
 }
